# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.FileControllerResponse:
    properties:
      file:
        $ref: '#/definitions/models.FileUploadResponse'
      files:
        $ref: '#/definitions/models.FileListResponse'
      message:
        type: string
      multi_files:
        $ref: '#/definitions/models.MultiFileUploadResponse'
    type: object
...
  models.FileListResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/models.FileUploadResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  models.FileUploadResponse:
    properties:
      created_at:
        type: string
      download_url:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: integer
      is_public:
        type: boolean
      mime_type:
        type: string
      original_name:
        type: string
    type: object
...
  models.MultiFileUploadResponse:
    properties:
      failed_files:
        items:
          type: string
        type: array
      failure_count:
        type: integer
      files:
        items:
          $ref: '#/definitions/models.FileUploadResponse'
        type: array
      success_count:
        type: integer
    type: object
...
  models.UpdateFileRequest:
    properties:
      is_public:
        type: boolean
    type: object
...
  utils.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/utils.APIError'
      message:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: JWT Authentication API with user registration and login
  title: Authentication API
  version: "1.0"
paths:
  /api/files/{id}:
    delete:
      description: Permanently delete a file owned by the authenticated user. This
        removes both the file record and the physical file from storage.
      parameters:
      - description: File ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: File deleted successfully
          schema:
            $ref: '#/definitions/controllers.FileControllerResponse'
        "400":
          description: Bad request - invalid file ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - you can only delete your own files
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to delete file
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete file
      tags:
      - File Management
    get:
      description: Retrieve detailed information about a specific file including metadata,
        download URL, and file properties
      parameters:
      - description: File ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: File information retrieved successfully
          schema:
            $ref: '#/definitions/controllers.FileControllerResponse'
        "400":
          description: Bad request - invalid file ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get file information by ID
      tags:
      - File Management
    put:
      consumes:
      - application/json
      description: Update file properties such as making it public or private. Only
        the file owner can update their files.
      parameters:
      - description: File ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: File update data - currently supports changing public/private
          status
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/models.UpdateFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: File updated successfully
          schema:
            $ref: '#/definitions/controllers.FileControllerResponse'
        "400":
          description: Bad request - invalid file ID or request body
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - you can only update your own files
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to update file
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update file properties
      tags:
      - File Management
  /api/files/my:
    get:
      description: Retrieve a paginated list of files uploaded by the authenticated
        user, including both public and private files
      parameters:
      - default: 1
        description: Page number for pagination
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User files retrieved successfully with pagination info
          schema:
            $ref: '#/definitions/controllers.FileControllerResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to retrieve files
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's uploaded files
      tags:
      - File Management
  /api/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Upload a single file for the authenticated user with comprehensive validation and metadata generation

        **File Requirements:**
        - Maximum size: 10MB per file
        - Supported formats: .jpg, .jpeg, .png, .gif, .pdf, .doc, .docx, .txt, .zip, .mp4, .mp3
        - Files are private by default (can be made public later via PUT /api/files/{id})

        **Upload Process:**
        1. File validation (size, type, format)
        2. Unique filename generation (timestamp + random hash)
        3. File storage in uploads directory
        4. Database record creation with metadata
        5. Automatic file type classification

        **Response includes:**
        - File ID and complete metadata
        - Download URL for accessing the file
        - Automatic file type classification (image, document, video, audio, archive, other)
        - Original filename and generated unique filename
        - File size, MIME type, and creation timestamp
      parameters:
      - description: 'File to upload (max 10MB). Use form field name ''file''. Example:
          curl -F ''file=@image.jpg'''
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: File uploaded successfully with metadata
          schema:
            $ref: '#/definitions/controllers.FileControllerResponse'
        "400":
          description: Bad request - file too large (>10MB), invalid type, or no file
            provided
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token in Authorization
            header
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to save file or create database
            record
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a single file
      tags:
      - File Upload
  /api/files/upload-multiple:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Upload multiple files in a single request. Each file is validated individually, allowing partial success scenarios.

        **Batch Upload Limits:**
        - Maximum 10 files per request
        - Total size limit: 50MB
        - Each file max: 10MB
        - Same format restrictions as single upload

        **Response includes:**
        - List of successfully uploaded files
        - Success/failure counts
        - Details of any failed uploads

        **Use Cases:**
        - Photo gallery uploads
        - Document batch processing
        - Media file collections
      parameters:
      - description: Multiple files to upload (use 'files' as field name)
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Files upload completed with detailed results
          schema:
            $ref: '#/definitions/controllers.FileControllerResponse'
        "400":
          description: Bad request - too many files, files too large, or no files
            provided
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to process files
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload multiple files
      tags:
      - File Upload
...
  /files/download/{filename}:
    get:
      description: Download a file by its unique filename. The file will be served
        with its original name and appropriate content type headers.
      parameters:
      - description: Unique filename (e.g., 1642678901_a1b2c3d4.jpg)
        in: path
        minLength: 1
        name: filename
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content with appropriate headers for download
          schema:
            type: file
        "400":
          description: Bad request - invalid or missing filename
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to serve file
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Download file
      tags:
      - File Download
  /files/public:
    get:
      description: Retrieve a paginated list of all files marked as public by users.
        No authentication required.
      parameters:
      - default: 1
        description: Page number for pagination
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Public files retrieved successfully
          schema:
            $ref: '#/definitions/controllers.FileControllerResponse'
        "500":
          description: Internal server error - failed to retrieve public files
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get public files
      tags:
      - File Browse
  /files/type/{type}:
    get:
      description: Retrieve public files filtered by file type. Useful for creating
        galleries or browsing specific content types.
      parameters:
      - description: File type filter
        enum:
        - image
        - document
        - video
        - audio
        - archive
        - other
        in: path
        name: type
        required: true
        type: string
      - default: 1
        description: Page number for pagination
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Files retrieved successfully by type
          schema:
            $ref: '#/definitions/controllers.FileControllerResponse'
        "400":
          description: 'Bad request - invalid file type (valid: image, document, video,
            audio, archive, other)'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to retrieve files
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get files by type
      tags:
      - Public/File Browse
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"